feat: Move models for identity (chrysalis) to api/chrysalis folder and fix imports
