+++ api/src/routes/chrysalis/identity/history/get.ts
@@ -2,12 +2,12 @@ import * as identityLegacy from "@iota/identity-wasm-0.4/node";
 import * as identity from "@iota/identity-wasm/node";
 
 import { ServiceFactory } from "../../../../factories/serviceFactory";
-import { IIdentityDidHistoryRequest } from "../../../../models/api/IIdentityDidHistoryRequest";
-import { IIdentityDidHistoryResponse } from "../../../../models/api/IIdentityDidHistoryResponse";
+import { IIdentityDidHistoryRequest } from "../../../../models/api/chrysalis/identity/IIdentityDidHistoryRequest";
+import { IIdentityDidHistoryResponse } from "../../../../models/api/chrysalis/identity/IIdentityDidHistoryResponse";
 import { IConfiguration } from "../../../../models/configuration/IConfiguration";
 import { CHRYSALIS } from "../../../../models/db/protocolVersion";
 import { NetworkService } from "../../../../services/networkService";
-import { IdentityHelper } from "../../../../utils/identityHelper";
+import { IdentityHelper } from "../../../../utils/chrysalis/identityHelper";
 import { ValidationHelper } from "../../../../utils/validationHelper";
 
 /**
