+++ api/src/routes/chrysalis/identity/resolution/get.ts
@@ -3,12 +3,12 @@ import * as identityLegacy from "@iota/identity-wasm-0.4/node";
 import * as identity from "@iota/identity-wasm/node";
 
 import { ServiceFactory } from "../../../../factories/serviceFactory";
-import { IIdentityDidResolveRequest } from "../../../../models/api/IIdentityDidResolveRequest";
-import { IIdentityDidResolveResponse } from "../../../../models/api/IIdentityDidResolveResponse";
+import { IIdentityDidResolveRequest } from "../../../../models/api/chrysalis/identity/IIdentityDidResolveRequest";
+import { IIdentityDidResolveResponse } from "../../../../models/api/chrysalis/identity/IIdentityDidResolveResponse";
 import { IConfiguration } from "../../../../models/configuration/IConfiguration";
 import { CHRYSALIS } from "../../../../models/db/protocolVersion";
 import { NetworkService } from "../../../../services/networkService";
-import { IdentityHelper } from "../../../../utils/identityHelper";
+import { IdentityHelper } from "../../../../utils/chrysalis/identityHelper";
 import { ValidationHelper } from "../../../../utils/validationHelper";
 
 /**
